[#233] Move profile requirements module to contrib.
