+++ apigee_devportal_kickstart.info.yml
@@ -66,7 +66,7 @@ install:
 - forum
 - paragraphs
 - pathauto
-- profile_requirement
+- requirements
 
 # List any themes that should be installed as part of the profile installation.
 # Note that this will not set any theme as the default theme.
