+++ composer.json
@@ -14,7 +14,8 @@
     "drupal/fontawesome": "^2.12",
     "drupal/paragraphs": "^1.6",
     "drupal/pathauto": "^1.3",
-    "drupal/radix": "^4.3"
+    "drupal/radix": "^4.3",
+    "drupal/requirements": "1.x-dev"
   },
   "config": {
     "sort-packages": true
