--- modules/custom/profile_requirement/phpcs.xml.dist
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-
-<ruleset name="Profile Requirement coding standards">
-  <exclude-pattern>*/.git/*</exclude-pattern>
-    <exclude-pattern>*/config/*</exclude-pattern>
-    <exclude-pattern>*/css/*</exclude-pattern>
-    <exclude-pattern>*/js/*</exclude-pattern>
-    <exclude-pattern>*/vendor/*</exclude-pattern>
-    <exclude-pattern>\.md</exclude-pattern>
-
-    <rule ref="Drupal"/>
-
-    <!-- Copyright header must be visible on all PHP files, including classes in a namespace. -->
-    <rule ref="Drupal.Commenting.FileComment.NamespaceNoFileDoc">
-      <severity>0</severity>
-    </rule>
-    <!-- Annotation classes still contains snake case variable names. -->
-    <rule ref="Drupal.NamingConventions.ValidVariableName.LowerCamelName">
-      <severity>0</severity>
-    </rule>
-    <!-- Core doesn't conform to this standard.
-         @see: https://www.drupal.org/project/coder/issues/2803251
-         @see: https://www.drupal.org/project/drupal/issues/1158720
-         -->
-    <rule ref="Drupal.Commenting.FunctionComment.TypeHintMissing">
-      <severity>0</severity>
-    </rule>
-</ruleset>
