--- modules/custom/profile_requirement/profile_requirement.info.yml
@@ -1,4 +1,0 @@
-name: Profile Requirement
-description: Provides a plugin for modules to implement profile requirements.
-type: module
-core: 8.x
