--- modules/custom/profile_requirement/profile_requirement.install
@@ -1,59 +1,0 @@
-<?php
-
-/**
- * @file
- * Copyright 2019 Google Inc.
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License version 2 as published by the
- * Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
- * License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- */
-
-use Drupal\Core\Link;
-use Drupal\Core\Url;
-
-/**
- * Implements hook_requirements().
- */
-function profile_requirement_requirements($phase) {
-  $requirements = [];
-
-  if ($phase === 'runtime') {
-    /** @var \Drupal\profile_requirement\Plugin\ProfileRequirementManagerInterface $profile_requirement_manager */
-    $profile_requirement_manager = Drupal::service('plugin.manager.profile_requirement');
-    $severity = REQUIREMENT_INFO;
-    $unmet_requirements = [];
-    /** @var \Drupal\profile_requirement\Plugin\ProfileRequirementInterface $requirement */
-    foreach ($profile_requirement_manager->listRequirements() as $requirement) {
-      if (!$requirement->isCompleted()) {
-        $unmet_requirements[$requirement->getId()] = $requirement->getLabel();
-        $severity = $requirement->getSeverity() === 'error' ? REQUIREMENT_ERROR : $severity;
-        $severity = $requirement->getSeverity() === 'warning' && $severity === REQUIREMENT_INFO ? REQUIREMENT_WARNING : $severity;
-      }
-    }
-
-    if (count($unmet_requirements)) {
-      $profile = _profile_requirement_get_install_profile();
-      $requirements['profile_requirement'] = [
-        'title' => $profile['name'],
-        'value' => Link::createFromRoute(t('Profile report'), 'profile_requirement.report'),
-        'description' => t('The following requirements for the installation profile have not been met: <strong>@requirements</strong>. <a href=":uri">See the profile report</a>.', [
-          '@requirements' => implode(', ', $unmet_requirements),
-          ':uri' => Url::fromRoute('profile_requirement.report')->toString(),
-        ]),
-        'severity' => $severity,
-      ];
-    }
-  }
-
-  return $requirements;
-}
