--- modules/custom/profile_requirement/profile_requirement.links.menu.yml
@@ -1,5 +1,0 @@
-profile_requirement.report:
-  title: 'Profile report'
-  parent: system.admin_reports
-  description: 'Displays a report of the installation profile''s status.'
-  route_name: profile_requirement.report
